gopls memory debug zip: in noname variant, replace names with hashes
To enable, provide a prefix to the '-elide' arg. Any time this prefix
is found, it will be replaced with 'elide' and the remaining path
elements will be replaced by hashes.

Hash input includes a runtime-generated nonce as well as path
elements preceding the current one - so should not leak anything
even the most security-conscious org would be concerned about.